--- src/components/NoSSR.tsx
@@ -1,8 +1,0 @@
-import dynamic from "next/dynamic";
-import React from "react";
-const NonSSRWrapper = (props: { children: any }) => (
-  <React.Fragment>{props.children}</React.Fragment>
-);
-export default dynamic(() => Promise.resolve(NonSSRWrapper), {
-  ssr: false,
-});
