--- src/pages/newindex.tsx
@@ -1,8 +1,0 @@
-import { Spawn } from "@/assets/maps";
-import { MapCanvas } from "@/components/MapCanvas";
-
-const Home = () => {
-  return <MapCanvas mapObject={Spawn} />;
-};
-
-export default Home;
