--- src/pages/teenindex.tsx
@@ -1,8 +1,0 @@
-import { Spawn } from "@/assets/maps";
-import { MapCanvas } from "@/components/MapCanvasOLD";
-
-const Home = () => {
-  return <MapCanvas mapObj={Spawn} />;
-};
-
-export default Home;
