--- src/utils/constants.ts
@@ -1,6 +1,0 @@
-export const CELL_WIDTH = 20;
-export const CELL_HEIGHT = 9;
-export const MAP_ID = "map";
-export const SELECTOR_ID = "selector";
-export const SPRITE_WIDTH = 16;
-export const SPRITE_HEIGHT = 16;
